requirements.txt
# バックエンド
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
requests==2.31.0

# フロントエンド
streamlit==1.28.0

# ユーティリティ
python-dateutil==2.8.2
loguru==0.7.2
